+++ backend/src/test/java/com/bd2_team6/biteright/user_preferences/UserPreferencesTests.java
@@ -99,8 +99,8 @@ public class UserPreferencesTests {
         user.setUserPreferences(preferences);
         userRepository.save(user);
 
-        Integer userId = user.getId();
-        Integer preferencesId = preferences.getUserPreferencesId();
+        Long userId = user.getId();
+        Long preferencesId = preferences.getUserPreferencesId();
 
         userRepository.deleteById(userId);
